# Make initial incremental/watch builds as fast normal builds