--- a/src/compiler/types.ts
+++ b/src/compiler/types.ts
@@ -5841,6 +5841,7 @@ namespace ts {
         declarationDir?: string;
         /* @internal */ diagnostics?: boolean;
         /* @internal */ extendedDiagnostics?: boolean;
+        /* @internal */ disableLazyShapeComputation?: boolean;
         disableSizeLimit?: boolean;
         disableSourceOfProjectReferenceRedirect?: boolean;
         disableSolutionSearching?: boolean;