@sokra 2021-02-26
Update:

* shape update propagation didn't work as intended. That's fixed now.
* fixed a bug where CommonJS modules were flagged as affecting the global scope. They no longer do that.
* I noticed that isolatedModules will affect the behavior, so I benchmarked that and added the results.