@sokra 2021-03-04
> [`lazySignatureCompute` (compare)](https://github.com/microsoft/TypeScript/compare/lazySignatureCompute) .. i just created commit with what meant when i said we can simplify this, obviously I haven't spent time looking at the test failures or change in deep to see if its correct and satisfies 0what you propose,

I see what you mean. That's probably a better way to solve that, compared to my added `NOT_COMPUTED_YET` values for the signature. With that we could use `undefined` as not computed value and skip the initial compilation with the additional flag you added.

> we will make this default without any option to disable this behavior

That's great. My PR currently adds a `disableLazyShapeComputation` option which is `/*@internal*/`. Won't that be enough or should we get rid of that too. In this case we might need to change some test cases a little bit.