@sheetalkamat 2021-03-04
> I guess I can integrate that into the PR.

Please do

> disableLazyShapeComputation

Please remove the option and fix test cases .. you may want to add more scenarios if they were testing initial local change, then add another local change as next step to ensure that still works correctly from then on 


Also can you please pull out deduplication part in separate PR and add test that shows it for easier review and maintenance. 

Thank you for great idea and work