@cnshenj 2021-03-05
There is one case not fully covered by the benchmark. In our project, we can easily repro the slow initial pass of watch/incremental. We also have another case that is very slow, may not be as slow as the initial pass, but slower than those benchmark test cases (those that take 8-9 seconds with or without the fix). The case is:
In VSCode, modify several files (I guess multiple files have shape affecting changes), then save them all with <kbd>Ctrl</kbd>+<kbd>K</kbd>, followed by <kbd>S</kbd>.