@sokra 2021-03-05
> There is one case not fully covered by the benchmark. In our project, we can easily repro the slow initial pass of watch/incremental. We also have another case that is very slow, may not be as slow as the initial pass, but slower than those benchmark test cases (those that take 8-9 seconds with or without the fix). The case is:
> In VSCode, modify several files (I guess multiple files have shape affecting changes), then save them all with Ctrl+K, followed by S.

That sounds like these files are referenced by many other files, which will all have their shape recomputed.

This PR tries to lazy compute shape to avoid to large cost on initial build. Once lazy computed, performance will be equal to eager computation.

Currently once a shape is computed it will keep computing it, to keep it up-to-date. Technically it's possible to throw the old shape away and go back to lazy mode. E. g. once the shapes that need to be computed grows over a certain threshold. But I guess that's something for a follow-up PR.