@sokra 2021-03-05
> [`lazySignatureCompute` (compare)](https://github.com/microsoft/TypeScript/compare/lazySignatureCompute) .. i just created commit with what meant when i said we can simplify this, obviously I haven't spent time looking at the test failures or change in deep to see if its correct and satisfies 0what you propose,

I made a few changes to this for the following reasons:

* Once a signature have been computed we can't ignore it as it would get out of date. So I renamed the flag to `avoidInitializingSignatures`
* When changing files or adding new files, this would eagerly compute the signature of them. That's fine when it's only a single file, but it would affect performance heavily when changing/adding many files.
  * e. g. switching branches
  * This also happens when the config file has errors that get fixed (see tests/baselines/reference/tsbuild/watchMode/noEmitOnError/does-not-emit-any-files-on-error-with-incremental.js). In that case there are a lot unprocessed files in `changedFilesSet` in the oldState.
  * I added a threshold for these cases (`> 3` changed/new files will avoid shape computation).