@sokra 2021-03-05
> Technically it's possible to throw the old shape away and go back to lazy mode. E. g. once the shapes that need to be computed grow over a certain threshold.

@cnshenj Actually this is quite easy to implement:

Could you try this branch? https://github.com/sokra/TypeScript/tree/performance/lazy-shape-fallback

It has this additional commit: https://github.com/sokra/TypeScript/commit/faee29ff8333ce0efb33aef1efa5124fce1a890b