@sokra 2021-03-05
> > I guess I can integrate that into the PR.
> 
> Please do

done

> 
> > disableLazyShapeComputation
> 
> Please remove the option and fix test cases .. you may want to add more scenarios if they were testing initial local change, then add another local change as next step to ensure that still works correctly from then on

It was hard, but done