@sokra 2021-03-07
> Also can you please pull out deduplication part in separate PR and add test that shows it for easier review and maintenance.

https://github.com/microsoft/TypeScript/pull/43079

> Should I also move [fix CommonJs modules no longer affecting the global scope `b7c2161` (#42960)](https://github.com/microsoft/TypeScript/pull/42960/commits/b7c2161f7157f6f6777b7837346d52b23bd61ee8) into a separate PR?

https://github.com/microsoft/TypeScript/pull/43090