@sokra 2021-03-12
> please dont bother updating the PR with conflicts

ok. It's a little bit hard to look at these conflicts and don't wanting to fix them...

Maybe you could look at the baselines of `when global file is added, the signatures are updated`? I think the test case may no longer test that signatures are updated, since they are not even computed with this change.