@sokra 2021-04-21
The latest beta version includes the fix